% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{arrange}
\alias{arrange}
\title{Arrange rows by variables}
\usage{
arrange(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Comma separated list of unquoted variable names, use `-` for
descending order.}
}
\value{
data.frame
}
\description{
Mimics dplyr::arrange using base R and rlang
}
\examples{
airquality\%>\%
arrange(-Month,Temp)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
