% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{b_arrange}
\alias{b_arrange}
\alias{b_arrange.default}
\alias{b_arrange.split_df}
\title{Arrange rows by variables}
\usage{
b_arrange(.data, ...)

\method{b_arrange}{default}(.data, ...)

\method{b_arrange}{split_df}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Comma separated list of unquoted variable names, use `-` for
descending order.}
}
\value{
data.frame
}
\description{
Mimics dplyr::arrange using base R and rlang
}
\examples{
airquality\%>\%
b_arrange(-Month,Temp)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
