% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{b_filter}
\alias{b_filter}
\alias{b_filter.default}
\alias{b_filter.split_df}
\title{Return rows with matching conditions}
\usage{
b_filter(.data, ...)

\method{b_filter}{default}(.data, ...)

\method{b_filter}{split_df}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Logical predicates defined in terms of the variables in .data}
}
\value{
data.frame
}
\description{
Mimics dplyr::filter using base R and rlang
}
\examples{
airquality\%>\%
b_filter(Temp > 80)\%>\%
head()

}
\seealso{
\code{\link[rlang]{quo_squash}},\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
