% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{b_rename}
\alias{b_rename}
\title{rename variables by name}
\usage{
b_rename(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{variables to rename}
}
\value{
data.frame
}
\description{
Mimics dplyr::rename using base R and rlang
}
\examples{
airquality\%>\%
b_rename(tmp = Temp,day = Day)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
