% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{b_select}
\alias{b_select}
\alias{b_select.default}
\alias{b_select.split_df}
\title{Select/rename variables by name}
\usage{
b_select(.data, ...)

\method{b_select}{default}(.data, ...)

\method{b_select}{split_df}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{unquoted variable names or character names}
}
\value{
data.frame
}
\description{
Mimics dplyr::select using base R and rlang
}
\examples{

airquality\%>\%
b_select(c(Day,Temp))\%>\%
head()

airquality\%>\%
b_select(c('Day','Temp'))\%>\%
head()

airquality\%>\%
b_select(-c(Day,Temp))\%>\%
head()

}
\seealso{
\code{\link[rlang]{quotation}},\code{\link[rlang]{quo_squash}}
}
\author{
Jonathan Sidi
}
