% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{b_ungroup}
\alias{b_ungroup}
\alias{b_ungroup.default}
\alias{b_ungroup.split_df}
\title{Remove grouping from data.frame}
\usage{
b_ungroup(.data)

\method{b_ungroup}{default}(.data)

\method{b_ungroup}{split_df}(.data)
}
\arguments{
\item{.data}{data.frame}
}
\value{
data.frame
}
\description{
Mimics dplyr::ungroup using base R
}
