% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\alias{count.default}
\alias{count.split_df}
\title{Count observations by group}
\usage{
count(x, ..., wt = NULL, sort = FALSE)

\method{count}{default}(x, ..., wt = NULL, sort = FALSE)

\method{count}{split_df}(x, ..., wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{...}{variables to group by}

\item{wt}{numeric, weights to count by, Default: NULL}

\item{sort}{boolean, if TRUE will sort output in descending order of n, default = FALSE}
}
\value{
data.frame
}
\description{
Mimics dplyr::count using base R and rlang
}
\examples{
airquality\%>\%
count(Month,Temp)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
