% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{filter}
\alias{filter}
\title{Return rows with matching conditions}
\usage{
filter(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Logical predicates defined in terms of the variables in .data}
}
\value{
data.frame
}
\description{
Mimics dplyr::filter using base R and rlang
}
\examples{
airquality\%>\%
filter(Temp > 80)\%>\%
head()

}
\seealso{
\code{\link[rlang]{quo_expr}},\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
