% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\alias{group_by.data.frame}
\title{FUNCTION_TITLE}
\usage{
group_by(.data, ..., add = FALSE)

\method{group_by}{data.frame}(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{add}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
