% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Count observations by group}
\usage{
rename(.data, ...)
}
\arguments{
\item{...}{variables to group by}

\item{x}{data.frame}

\item{wt}{numeric, weights to count by, Default: NULL}

\item{sort}{boolean, if TRUE will sort output in descending order of n, default = FALSE}
}
\value{
data.frame
}
\description{
Mimics dplyr::count using base R and rlang
}
\examples{
airquality\%>\%
rename(tmp = Temp,day = Day)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
