% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select.default}
\alias{select.split_df}
\title{Select/rename variables by name}
\usage{
select(.data, ...)

\method{select}{default}(.data, ...)

\method{select}{split_df}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{unquoted variable names or character names}
}
\value{
data.frame
}
\description{
Mimics dplyr::select using base R and rlang
}
\examples{

airquality\%>\%
select(c(Day,Temp))\%>\%
head()

airquality\%>\%
select(c('Day','Temp'))\%>\%
head()

airquality\%>\%
select(-c(Day,Temp))\%>\%
head()

}
\seealso{
\code{\link[rlang]{quotation}},\code{\link[rlang]{quo_expr}}
}
\author{
Jonathan Sidi
}
