% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\name{spread}
\alias{spread}
\title{base spread}
\usage{
spread(data, key, value, convert = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{key}{column which will become the new columns}

\item{value}{column name which will populate new columns}

\item{convert}{type.convert will run on each result column if TRUE, Default: FALSE}
}
\value{
data.frame
}
\description{
base spread mimics basic functionality of tidyr::spread
}
\examples{
stocks <- data.frame(
time = as.Date('2009-01-01') + 0:9,
X = rnorm(10, 0, 1),
Y = rnorm(10, 0, 2),
Z = rnorm(10, 0, 4)
)
stocksm <- gather(stocks,'stock', 'price', -1)
spread(stocksm, 'stock', 'price')
# spread and gather are complements
df <- data.frame(x = c("a", "b"), y = c(3, 4), z = c(5, 6))
sdf <- spread(df, 'x', 'y')
gather(sdf, 'x', 'y', -1, na.rm = TRUE)

# Use 'convert = TRUE' to produce variables of mixed type
df <- data.frame(row = rep(c(1, 51), each = 3),
                 var = c("Sepal.Length", "Species", "Species_num"),
                 value = c(5.1, "setosa", 1, 7.0, "versicolor", 2))

str(spread(df, 'var', 'value'))
str(spread(df, 'var', 'value',convert = TRUE))
}
\author{
Jonathan Sidi
}
