% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{ungroup}
\alias{ungroup}
\alias{ungroup.default}
\alias{ungroup.split_df}
\title{Remove grouping from data.frame}
\usage{
ungroup(.data)

\method{ungroup}{default}(.data)

\method{ungroup}{split_df}(.data)
}
\arguments{
\item{.data}{data.frame}
}
\value{
data.frame
}
\description{
Mimics dplyr::ungroup using base R
}
