% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{unite}
\alias{unite}
\title{Unite multiple columns into one.}
\usage{
unite(data, col, ..., sep = "_", remove = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{col}{character, name of the new column}

\item{\dots}{columns to combine}

\item{sep}{character, separator to use between values, Default: '_'}

\item{remove}{boolean, if TRUE remove input columns from output object, Default: TRUE}
}
\value{
data.frame
}
\description{
Mimics tidyr::unite using base R and rlang
}
\details{
the main difference between this lite version and the tidyr version is that the
new column is attached to the end of the data.frame and not before the index of the first column
that is to be united. Since this is mainly aesthetic it was not transfered over.
}
\examples{

unite(airquality, col = "month_day", columns = c(Month,Day))\%>\%
head

unite(airquality, col = "temp_month_day", columns = c(Temp:Day))\%>\%
head

unite(airquality, col = "month_day", columns = c("Month","Day"))\%>\%
head


}
\author{
Jonathan Sidi
}
