% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\alias{group_by.data.frame}
\title{Group by one or more variables}
\usage{
group_by(.data, ..., add = FALSE)

\method{group_by}{data.frame}(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Variables to group by}

\item{add}{When add = FALSE, the default, group_by() will override existing groups.
To add to the existing groups, use add = TRUE., Default: FALSE}
}
\value{
data.frame
}
\description{
Mimics dplyr::group_by using base R and rlang
}
