% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\alias{mutate.default}
\alias{mutate.split_df}
\title{Add new variables}
\usage{
mutate(.data, ...)

\method{mutate}{default}(.data, ...)

\method{mutate}{split_df}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{Name-value pairs of expressions}
}
\value{
data.frame
}
\description{
Mimics dplyr::mutate using base R and rlang
}
\examples{
airquality\%>\%
mutate(
    lOzone = log(Ozone),
    Month  = factor(month.abb[Month]),
    cTemp  = round((Temp - 32) * 5/9, 1),
    S.cT   = Solar.R / cTemp
)\%>\%
head()
}
\seealso{
\code{\link[rlang]{quotation}}
}
\author{
Jonathan Sidi
}
