% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{ungroup}
\alias{ungroup}
\alias{ungroup.default}
\alias{ungroup.split_df}
\title{FUNCTION_TITLE}
\usage{
ungroup(.data)

\method{ungroup}{default}(.data)

\method{ungroup}{split_df}(.data)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
