% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\name{unite}
\alias{unite}
\title{base unite}
\usage{
unite(data, col, columns = NULL, regex = FALSE, sep = "_",
  remove = TRUE, ...)
}
\arguments{
\item{data}{data.frame}

\item{col}{character, name of the new column}

\item{columns}{column names or indicies or regex of them to gather,
Default: NULL}

\item{regex}{boolean, indicates of columns is to be treated as a
regular expression, Default: FALSE}

\item{sep}{character, separator to use between values, Default: '_'}

\item{remove}{boolean, if TRUE remove input columns from output object, Default: TRUE}

\item{\dots}{parameters to pass to grep}
}
\value{
data.frame
}
\description{
base unite mimics basic functionality of tidyr::unite
}
\details{
the main difference between this lite version and the tidyr version is that the
new column is attached to the end of the data.frame and not before the index of the first column
that is to be united. Since this is mainly aesthetic it was not transfered over.
}
\examples{

unite(mtcars, col = "vs_am", columns = c("vs","am"))

unite(mtcars, col = "disp_drat", columns = '^d', regex = TRUE)

}
\author{
Jonathan Sidi
}
